Rename: reorganized packages
패키지 구조 통일하여 수정하였습니다.
