+++ src/main/java/com/example/stagealarm/artist/entity/ArtistLike.java
@@ -1,24 +1,25 @@
-package com.example.stagealarm.artist;
+package com.example.stagealarm.artist.entity;
 
 import com.example.stagealarm.BaseEntity;
-import com.example.stagealarm.genre.entity.Genre;
+import com.example.stagealarm.artist.entity.Artist;
+import com.example.stagealarm.user.entity.UserEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.ManyToOne;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 
 @Getter
 @Builder
 @Entity
 @NoArgsConstructor
 @AllArgsConstructor
-public class ArtistGenre extends BaseEntity {
+public class ArtistLike extends BaseEntity {
 
   @ManyToOne(fetch = FetchType.LAZY)
+  @Setter
   private Artist artist;
   @ManyToOne(fetch = FetchType.LAZY)
-  private Genre genre;
+  @Setter
+  private UserEntity userEntity;
+
 }
