+++ src/main/java/com/example/stagealarm/artist/repo/ArtistLikeRepository.java
@@ -1,8 +1,8 @@
 package com.example.stagealarm.artist.repo;
 
-import com.example.stagealarm.artist.Artist;
-import com.example.stagealarm.artist.ArtistLike;
-import com.example.stagealarm.user.UserEntity;
+import com.example.stagealarm.artist.entity.Artist;
+import com.example.stagealarm.artist.entity.ArtistLike;
+import com.example.stagealarm.user.entity.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
