+++ src/main/java/com/example/stagealarm/artist/repo/ArtistRepository.java
@@ -1,6 +1,6 @@
 package com.example.stagealarm.artist.repo;
 
-import com.example.stagealarm.artist.Artist;
+import com.example.stagealarm.artist.entity.Artist;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ArtistRepository extends JpaRepository<Artist, Long> {
