+++ src/main/java/com/example/stagealarm/artist/service/ArtistLikeService.java
@@ -1,11 +1,11 @@
 package com.example.stagealarm.artist.service;
 
-import com.example.stagealarm.artist.Artist;
-import com.example.stagealarm.artist.ArtistLike;
+import com.example.stagealarm.artist.entity.Artist;
+import com.example.stagealarm.artist.entity.ArtistLike;
 import com.example.stagealarm.artist.dto.ArtistLikeDto;
 import com.example.stagealarm.artist.repo.ArtistLikeRepository;
 import com.example.stagealarm.artist.repo.ArtistRepository;
-import com.example.stagealarm.user.UserEntity;
+import com.example.stagealarm.user.entity.UserEntity;
 import com.example.stagealarm.user.repo.UserRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
