+++ src/main/java/com/example/stagealarm/genre/entity/GenreSubscribe.java
@@ -1,8 +1,7 @@
 package com.example.stagealarm.genre.entity;
 
 import com.example.stagealarm.BaseEntity;
-import com.example.stagealarm.genre.entity.Genre;
-import com.example.stagealarm.user.UserEntity;
+import com.example.stagealarm.user.entity.UserEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
 import jakarta.persistence.ManyToOne;
